MVP - Renames WFOStepList folder to WFOWorkflowStepList
