+++ packages/orchestrator-ui-components/src/components/WFOSteps/index.ts
@@ -1,2 +1,2 @@
 export * from './WFOStep';
-export * from './WFOStepList';
+export * from './WFOWorkflowStepList';
